+++ src/components/Company/Company.js
