+++ src/components/Company/EditCompany.js
