+++ src/components/Computer/Computer.jsx
