+++ src/containers/company/Companies.hook.js
@@ -1,4 +1,4 @@
-import  {MOCK_COMPANIS}  from "./Mock";
+import  {MOCK_COMPANIS}  from "../../Mock";
 export function useCompanies(){
     return(MOCK_COMPANIS);
 }
