+++ src/containers/computer/Computers.hook.js
@@ -1,4 +1,4 @@
-import {MOCK_COMPUTERS} from './Mock'
+import {MOCK_COMPUTERS} from '../../Mock'
 
 export function useComputers(){
     return(MOCK_COMPUTERS.list_computer)
